+++ src/components/add-remove/add-remove.spec.ts
@@ -1,0 +1,14 @@
+import AddRemoveElement from './add-remove';
+import { createElementFromString } from '../../common/WebComponent';
+
+test('Check rendering authors text', () => {
+  const addremove = createElementFromString(`
+    <div class="switch-wrapper">
+      <button class="switch switch--add">+</button>
+      <button class="switch switch--remove">-</button>
+    </div>`);
+
+  const addRemoveElement = new AddRemoveElement('+', '-');
+
+  expect(addRemoveElement.render()).toStrictEqual(addremove);
+});